Add check for parent UI Figure
Previous logic assumed that the parent of the input UI object is the
UI Figure, which is an incorrect assumption. Other containers, like
uipanel, exist and are valid parents for other UI objects but will
cause the object type check to fail.

This type check has been updated to utilize ancestor to obtain the
parent figure window.

See #6